chore: clean plugin dist

+++ .erb/scripts/clean.js
@@ -1,6 +1,8 @@
 import rimraf from 'rimraf';
-import webpackPaths from '../configs/webpack.paths.ts';
+import webpackPaths from '../configs/webpack.paths';
 import process from 'process';
+import path from 'path';
+import fs from 'fs';
 
 const args = process.argv.slice(2);
 const commandMap = {
@@ -8,11 +10,20 @@ const commandMap = {
   release: webpackPaths.releasePath,
   dll: webpackPaths.dllPath,
   build: webpackPaths.buildPath,
+  plugin: webpackPaths.pluginPath,
 };
 
 args.forEach((x) => {
   const pathToRemove = commandMap[x];
-  if (pathToRemove !== undefined) {
+  if (pathToRemove !== undefined && pathToRemove !== commandMap.plugin) {
     rimraf.sync(pathToRemove);
+  }
+
+  if(pathToRemove === commandMap.plugin) {
+    fs.readdirSync(pathToRemove, { withFileTypes: true }).forEach((plugin) => {
+      if(plugin.isDirectory()) {
+        rimraf.sync(path.join(pathToRemove, plugin.name, 'dist'));
+      }
+    })
   }
 });
